+++ src/BuildSelectSqlCapableTrait.php
@@ -26,15 +26,28 @@ trait BuildSelectSqlCapableTrait
      *
      * @since [*next-version*]
      *
-     * @param array|stdClass|Traversable        $columns  The columns, as a map of aliases (as keys) mapping to
-     *                                                    column names, expressions or entity field instances.
-     * @param array|stdClass|Traversable        $tables   A mapping of tables aliases (keys) to their real names.
-     * @param array|Traversable                 $joins    A list of JOIN logical expressions, keyed by table name.
-     * @param LogicalExpressionInterface|null   $where    The WHERE logical expression condition.
-     * @param OrderInterface[]|Traversable|null $ordering The ordering, as a list of OrderInterface instances.
-     * @param int|null                          $limit    The number of records to limit the query to.
-     * @param int|null                          $offset   The number of records to offset by, zero-based.
-     * @param array                             $hashmap  Optional map of value names and their hashes.
+     * @param array|stdClass|Traversable                                        $columns  The columns, as a map of
+     *                                                                                    aliases (as keys) mapping to
+     *                                                                                    column names, expressions or
+     *                                                                                    entity field instances.
+     * @param array|stdClass|Traversable                                        $tables   A mapping of tables aliases
+     *                                                                                    (keys) to their real names.
+     * @param array|Traversable                                                 $joins    A list of JOIN logical
+     *                                                                                    expressions, keyed by table
+     *                                                                                    name.
+     * @param LogicalExpressionInterface|null                                   $where    The WHERE logical expression
+     *                                                                                    condition.
+     * @param OrderInterface[]|Traversable|null                                 $ordering The ordering, as a list of
+     *                                                                                    OrderInterface instances.
+     * @param int|null                                                          $limit    The number of records to
+     *                                                                                    limit the query to.
+     * @param int|null                                                          $offset   The number of records to
+     *                                                                                    offset by, zero-based.
+     * @param string[]|Stringable[]|EntityFieldInterface[]|stdClass|Traversable $grouping A list of strings, stringable
+     *                                                                                    objects or entity-field
+     *                                                                                    instances.
+     * @param array                                                             $hashmap  Optional map of value names
+     *                                                                                    and their hashes.
      *
      * @throws InvalidArgumentException If an argument is invalid.
      * @throws OutOfRangeException      If the limit or offset are invalid numbers.
@@ -49,6 +62,7 @@ trait BuildSelectSqlCapableTrait
         $ordering = null,
         $limit = null,
         $offset = null,
+        $grouping = [],
         array $hashmap = []
     ) {
         if ($this->_countIterable($tables) === 0) {
@@ -64,6 +78,7 @@ trait BuildSelectSqlCapableTrait
         $from = $this->_buildSqlFrom($tables);
         $rJoins = $this->_buildSqlJoins($joins, $hashmap);
         $rWhere = $this->_buildSqlWhereClause($where, $hashmap);
+        $rGroup = $this->_buildSqlGroupByClause($grouping);
 
         $sOrder = ($ordering !== null)
             ? $this->_buildSqlOrderBy($ordering)
@@ -75,7 +90,7 @@ trait BuildSelectSqlCapableTrait
             ? $this->_buildSqlOffset($offset)
             : '';
 
-        $parts = array_filter([$from, $rJoins, $rWhere, $sOrder, $sLimit, $sOffset], 'strlen');
+        $parts = array_filter([$from, $rJoins, $rWhere, $rGroup, $sOrder, $sLimit, $sOffset], 'strlen');
         $tail = implode(' ', $parts);
         $query = sprintf(
             'SELECT %1$s %2$s;',
@@ -181,6 +196,23 @@ trait BuildSelectSqlCapableTrait
      * @return string The built OFFSET query portion.
      */
     abstract protected function _buildSqlOffset($offset = null);
+
+    /**
+     * Builds the GROUP BY portion of the an SQL query.
+     *
+     * @since [*next-version*]
+     *
+     * @param string[]|Stringable[]|EntityFieldInterface[]|stdClass|Traversable $grouping A list of strings, stringable
+     *                                                                                    objects or entity-field
+     *                                                                                    instances.
+     *
+     * @return string The built GROUP BY query portion.
+     *
+     * @throws InvalidArgumentException   If the argument is not a valid iterable.
+     * @throws OutOfRangeException        If an element in the iterable is invalid.
+     * @throws InternalExceptionInterface If a problem occurred while trying to retrieve a column name.
+     */
+    abstract protected function _buildSqlGroupByClause($grouping = []);
 
     /**
      * Counts the elements in an iterable.